Close #188 - [extras-scala-io] Add syntax: Seq[String].rainbowed
