+++ modules/extras-scala-io/shared/src/main/scala-2/extras/scala/io/syntax/truecolor/RainbowSyntax.scala
@@ -1,6 +1,6 @@
 package extras.scala.io.syntax.truecolor
 
-import extras.scala.io.syntax.truecolor.RainbowSyntax.RainbowSyntaxOps
+import extras.scala.io.syntax.truecolor.RainbowSyntax.{RainbowSyntaxOps, RainbowsSyntaxOps}
 import extras.scala.io.truecolor.Rainbow
 
 /** @author Kevin Lee
@@ -8,10 +8,17 @@ import extras.scala.io.truecolor.Rainbow
   */
 trait RainbowSyntax {
   implicit def rainbowSyntax(s: String): RainbowSyntaxOps = new RainbowSyntaxOps(s)
+
+  implicit def rainbowsSyntax(ss: Seq[String]): RainbowsSyntaxOps = new RainbowsSyntaxOps(ss)
 }
 object RainbowSyntax {
   class RainbowSyntaxOps(private val s: String) extends AnyVal {
     def rainbowed: String     = Rainbow.rainbow(s)
     def rainbowedHtml: String = Rainbow.rainbowHtml(s)
   }
+
+  class RainbowsSyntaxOps(private val ss: Seq[String]) extends AnyVal {
+    def rainbowed: Seq[String] = Rainbow.rainbows(ss)
+  }
+
 }
