+++ modules/extras-scala-io/shared/src/main/scala-3/extras/scala/io/syntax/truecolor/RainbowSyntax.scala
@@ -10,4 +10,8 @@ trait RainbowSyntax {
     def rainbowed: String     = Rainbow.rainbow(s)
     def rainbowedHtml: String = Rainbow.rainbowHtml(s)
   }
+
+  extension (ss: Seq[String]) {
+    def rainbowed: Seq[String] = Rainbow.rainbows(ss)
+  }
 }
